requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
django-cors-headers>=4.3
django-modeltranslation>=0.18
requests>=2.31
openai>=1.0
google-generativeai>=0.3.0
# xAI Grok использует OpenAI-совместимый API, поэтому xai-sdk не требуется (требует Python >= 3.10)
